vaapiencode: h264: set best profile in the intersected caps
The previous behavior when selecting a profile was to select the
best one in the intersection of allowed, in the pipelne, and
available (by the hardware). This was changed in commit 1f1e9ee
to choose the first one at fixate.

Nonetheless, it's a bad idea change the behavior on a stable
release. Thus, this patch selects the best profile in the
intersected caps as before.