uninstalled: Fix meson detection when using meson from a git checkout
This would not trigger on the build bot because there is
special casing in the code for a meson checkout underneath
gst-build. Fix needed as mesonintrospect.py was changed into
'meson.py introspect' in recent meson versions. When using
meson from git to configure the build the uninstalled script
would pick up a system meson instead which then would then
error out parsing the coredata from the newer meson.